refactor(collections): extract the smart rules editor into one component
The condition editor was duplicated verbatim across the create and edit
dialogs, so every fix had to be made twice.
